Update jparse.md
Much like aba06a5f5f2b42adde25f332a44513cfd6d2c264 I have removed from
the bugs section that it does not print the filename as it now does
print filenames. It might be better to say that in some cases it does
not but generally it does but this will have to come at another time as
I really must leave now!

I also updated the syntax and an example to use -- to separate options
from the rest of the command line. This is not done in the man pages yet
for a technicality with run_usage.sh but it will be done in time (for
EXAMPLES it should be done though but again this will happen later).