Fix de1561a3965a14520815025c7a8953e4e843a09e
There were two problems to be worked out. The first and easy one was I
removed the shadowing of the filename variable. In the function it's
just 'name'.

The second one was more obscure. If one did make parser all everything
would compile okay. If however one were to do:

    ./run_bison.sh
    make

one would get a compilation error with something like:

    jparse.y:73:7: error: typedef redefinition with different types ('struct YYLTYPE' vs 'struct YYLTYPE')
	} YYLTYPE;
	  ^
    jparse.y:73:7: note: previous definition is here
	} YYLTYPE;
	  ^
    1 error generated.
    make: *** [jparse.tab.o] Error 1

This is odd because make parser indirectly runs ./run_bison.sh itself.

The solution is that it needed an #ifdef guard in the code requires
block in jparse.y. The best I can figure is because jparse.h includes
jparse.tab.h and jparse.tab.h is where YYLTYPE is redefined that
somewhere in the chain it was included a second time which made it seem
like it was redefining it with the same thing (for your eyes do not
deceive you: those were the exact same things and locations in the
file). To solve the problem I simply added an #if
!defined(JPARSE_TAB_H_INCLUDED) to jparse.h (not sure if that is needed
as it didn't solve the problem but I forgot about it) and also an #if
!defined(YYLTYPE_IS_DECLARED) to jparse.y where only then would the type
be redeclared. It was the latter one that solved the problem but since
the other one does not cause a problem I'm keeping it in place.