Mass man pages fixes
The most common fix is that in the SEE ALSO section apparently one
should not end the lines with a full stop. There are some exceptions in
this repo namely when the lines in the section are sentences.

I also changed it to be more proper: '..., ..., ...' to '..., ... and
...'.

I removed empty lines (some were left in dbg.3 because those are in a C
source file). Since dbg.3 was updated the dbg repo also needs to be
updated but I will do that later on. If I do not get to it it should not
matter as it eventually will be done but this is kind of a note to
myself as well. The reason this was done is it seems that this is the
proper style - whatever proper might mean.

There might be a few other changes as well but the above changes are the
main ones.