Comment fix in txzchk.c and utf8_posix_map.c
For txzchk.c I made the comment for inclusion of strings.h consistent
with the other source files that include it. Specifically they just have
the function name but txzchk.c had ' for ' in front of the function
name.

In utf8_posix_map.c there was no comment for strings.h.

The header file is not necessary in some systems but is in others so
it's useful to say why it's needed. In the case of txzchk (and some
others) it's for strcasecmp(). For utf8_posix_map.c it's for
strncasecmp().