Minor update to c5daa500d51b9
Just in case we now have the YY_USER_ACTION macro also use the getter
functions.

I have updated the jparse.md and jparse.1 files to note that although
the scanner is re-entrant also it is not tested yet because it's out of
scope of this repo. Later on it can be tested better. Our tools and test
scripts work fine of course.