Update run_flex.sh for jparse.lex.h
Other files updated: Makefile (including make depend) plus the man page.

There were some typo fixes as well.

The jparse.y now #includes "jparse.lex.h" instead of "jparse.lex.ref.h"
as the script run_flex.sh now works.

With this all that remains is to address the changing the filename to be
local and possibly the same thing for the number of errors (if we even
need that still).