Don't reference yyin in jparse.l
When I made the scanner re-entrant I commented out the assignment of
yyin to the stream but I forgot to remove the comment about it and that
comment itself. In order to get or set yyin one needs the scanner data
but the function that parses the json data is not the one that has the
FILE * so there's no need to use it at all.