Fix segfault in some systems introduced in c5daa500
Under macOS this problem did not show up but under linux it did. The
problem is that chkentry also called clearerr_or_fclose() on the streams
of the .info.json and .auth.json files but because the scanner also uses
this it resulted in a double free. I'm not sure exactly why this did not
happen before since it was always the same stream in the end but
possibly because yyin was used it was in a different state. As I noted
in either a commit log or a GitHub comment we can no longer use yyin due
to complexities introduced with the re-entrant scanner wrt
setting/getting yyin needing the scanner context (which we don't always
have): specifically the function that opens the json stream (or stream
that is supposed to be json) does not have the scanner. Instead the
function that that function calls has the scanner context so we can't
use yyin. We obviously have to close it in the scanner since at that
point it'll always be reached but we don't always use chkentry so there
would be a file descriptor leak.