Fix memory leak in jparse
In main() in jparse_main.c the ordering of some of the code would result
in a memory leak. Specifically if the JSON was invalid then tree would
never be freed because it exited before this (though the OS should
handle that). Even more the call to json_tree_free() was only called if
tree was NULL.

There also was an ordering issue with reporting if tree was NULL
(because the tree is NULL the json is invalid and in this case the
program was exited before the warning could be issued - as well as the
freeing of the tree if not NULL).