JSON scanner more re-entrant (filename local)
With this commit the filename is no longer global. This means the only
thing left that is global (besides the yydebug but that is always global
and we have no control over that) is the num_errors and this might be a
variable that will be removed as it's not even clear it's useful now.