Minor fixes and comment clarification in jparse.l
In the YY_USER_ACTION we check that yyextra != NULL before dereferencing
it (for filename). It should never be NULL anyway but we check
nonetheless. Maybe similar could be done for yylloc but what the values
would be to use is not so clear (for filename it's just "").

Fix comment about yy_delete_buffer().