Isolate use of <time.h>, remove D_XOPEN_SOURCE from CFLAGS
We isolate the use of <time.h> and <sys/time.h> to a entry_time.c
so that the use of `#define _XOPEN_SOURCE` can be restricted so
that the `_XOPEN_SOURCE` does not impact in unexpected ways, other
libc functions.

We removed `-D_XOPEN_SOURCE=600` from the Makefile variable
`$(D_LEGACY}` and thus it is not part of the C compile Makefile
variable `${CFLAGS}`.

We removed the need for `TIMEGM_PROBLEM` and so it was removed from
the Makefile variable `$(D_LEGACY}` and thus it is not part of the
C compile Makefile variable `${CFLAGS}`.

We adjusted the use of include files in `entry_time.c`,
`entry_time.h` and used include files in the new `entry_time.c`
and `entry_time.h` along the following concept:

* Only include files in a _.h_ header file that that are required
  in order to include the given header file.

* Only includes files in a _.c_ source file that are required
  to compile the _.c_ source file.