Bug fixes and improvements in several scripts
Several scripts under test did not show the version right in the case of
bug_report.sh. Specifically I removed the 1>&2 from the echo of the
version as with that there it showed a different and confusing output.
For example:

    ## Checking if ./test/txzchk_test.sh is executable
    ## ./test/txzchk_test.sh IS EXECUTABLE
    ## RUNNING: ./test/txzchk_test.sh -V
    0.5 2022-11-04
    ./test/txzchk_test.sh version

Notice the last two lines where the version by itself is printed on a
line and then below it it just says the name of the script followed by
version followed by a space and then newline. It now looks like:

    ## Checking if ./test/txzchk_test.sh is executable
    ## ./test/txzchk_test.sh IS EXECUTABLE
    ## RUNNING: ./test/txzchk_test.sh -V
    ./test/txzchk_test.sh version 0.5 2022-11-04

which is cleaner and clearer.

Updated bug_report.sh version. It now has a new function
get_version_minimal which tries to get the version of a tool in a
minimal way. Consider cat(1). Using -v is problematic because it blocks
(it's an option and then expects input from stdin). -V might or might
not exist. --version also might or might not exist. What we try on it is
--version and then also (if available) what(1) and (if available)
ident(1). If it does not find anything beyond that it returns and does
not even warn that there's an unknown version as it's almost certainly
not important and would just clutter the log (we don't use strings(1) on
it for that reason as well).

The bug_report.sh script also has had some clarifying comments (and some
humour in comments as well) added. There possibly were other things with
the script.

The ioccc_test.sh did not use -Z option for jstr_test.sh which caused
some problems. More importantly jstr_test.sh itself was fixed so that
raw references (for when needing the filename itself to cat) of
jstr_test.sh were changed to $0. This fixes a problem I reported (#407)
but there still is another problem with that script.

There possibly were other changes but these are the important ones if
any others were in fact made.