Remove global num_errors from jparse
As of this commit with the exception of the one we have no control over
(yydebug) there are no global variables left with the scanner or parser!

There is one change with this commit that can result in an error
condition (or theoretically could) and that's if a NULL is_valid bool is
passed to the parse functions. We need that to indicate whether the json
is valid or invalid.

As for why we couldn't do it in extra data like the filename. It might
have been possible but it would cause extra complications by having to
pass an additional pointer to a struct (struct json_extra *) to various
functions and this seems not worth it since we don't really need the
num_errors variable anyway any more (we used to but it's no longer
useful). If however this changes it certainly is possible to add it back
but in a re-entrant way.