Fix bug in jstr_test.sh - resolves #407
Now the script checks for failure of test 3. I'm not sure if there are
any conditions that might not be picked up. For example if another part
of the pipeline fails it might not be detected but at least if one of
the files does not exist (or cannot be found) the test will fail.