Improve 8b2ed9bbd4f4b16844171e9d6e5c90df7ee7879f
The script jstr_test.sh now checks the status of each part of the
pipeline in test #3. A point of interest is that it does not say what
part of the test failed: it just shows that the test failed. This might
or might not be something that should be changed.