Typo fixes in reset_tstamp.sh
In addition to changing an actual typo I have made the formatting of the
help string consistent across all lines (specifically the -l option had
the second column further to the right compared to the other options). I
also changed the error messages related to the limit file a bit to match
the option change itself (the option argument was changed from
limit_ioccc.h - which is the default - to limit.h).