README.md fixes
After the man pages were moved to the man/ subdirectory the example
commands to get the man pages were of course no longer correct.

I also elaborated a bit more on txzchk (more specifically it now matches
txzchk.{c,h}). This was done because it will now be seen on the main
page of the repo at GitHub which seems like an important place to have
it (it is short).