Typo fixes in verge.8
Well actually one typo fix and change the >= to use the \(>= escape
sequence (there are other man pages that this should also probably do
this).

The typo was in the EXAMPLES section to do with the meaning of the exit
value. Specifically it suggested the example checks if the first version
is > the second version but it's actually >=. I used the \(>= escape
sequence there too (which on my system shows as just >= but it might
show up differently on other systems .. whether that's a problem or not
I don't know but since it's meant to be for this purpose it seems more
correct).