Restructure dbg facility and Makefile
This is a major update which probably can be and if so should be
improved (as part of it feels hackish) but it's an important first step
to issue #87 Pre-release 1.0 actions.

The dbg facility was moved over to dbg/ subdirectory. This was done via
some rules that were recently added to the Makefile. Now when the dbg
repo is updated we should be able to just run a rule and the dbg
facility in this repo will be synchronised. This will likely be done
with the dyn_array facility and later on the JSON parser as well.

Besides changing all '#include "dbg.h"' to '#include "dbg/dbg.h"' this
did involve quite a few changes in the Makefile as well. The main thing
is that references to dbg related files now are in dbg/ subdirectory
with the exception of in the dependencies (dbg.o: line at the end of the
Makefile). That latter part might or might not need to be changed in the
future.

One of the things that feels like a hack (at least to me) is that
instead of using $@ in all cases sometimes we have to use ../$@ because
we switched to the dbg subdirectory which means that $@ would no longer
be valid.

I also made an important fix to do with changing directories that also
has some safety implications. Specifically lines in the form of

     cd foo; ...

were changed to

    cd foo && ...

This is important because if the directory does not exist it can cause a
lot of problems depending on the command that follows. In the case of
the dbg cloning rules this is probably more important but even the cd
test commands are better if we only do the command if the change in
directory succeeds (imagine if the test/ directory was removed or is
missing for some reason).

Now as far as what can be done to improve this: the subdirectories
probably need their own Makefiles. However with dbg there already exists
a Makefile so it might need to be modified in some way to work with this
repo or else a new Makefile (in addition - not replacing) should be
added. Whether that's added to just this repo or the other is not yet
determined. The problem with having it just in this repo is that dbg/ is
removed - albeit temporarily - in some rules (maybe that's a problem,
maybe it's not) so it might be that the new Makefile (maybe Makefile.dbg
?) will have to be in the main directory but this is also to be
determined. One thing that might be useful is to make a dbg rule in the
main Makefile which does some of the work of the dbg related rules but
this will have to be done at another time as well. This is also yet to
be determined.