Improve Makefile dbg rules
Remove the following explicit rules from the Makefile:

    dbg/dbg.o: dbg/dbg.c dbg/dbg.h Makefile
	   cd dbg && ${CC} ${CFLAGS} dbg.c -c

    dbg/dbg_test.c: dbg/dbg.c Makefile
	   cd dbg && ${RM} -f $@; ${CP} -v -f dbg.c ../$@

    dbg/dbg_test.o: dbg/dbg_test.c dbg/dbg.h Makefile
	   cd dbg && ${CC} ${CFLAGS} -I.. -DDBG_TEST dbg_test.c -c

    dbg/dbg_test: dbg/dbg_test.o Makefile
	   cd dbg && ${CC} ${CFLAGS} -I.. dbg_test.o -o ../$@

In their place there is a dbg rule that just does:

	${MAKE} -C dbg

which lets the dbg/Makefile handle everything because instead of
requiring dbg/dbg.o (for example) it's just dbg which forces the dbg
rule to run.

Additionally the dbg rule copies dbg/dbg.3 to man/dbg.3. The man/dbg.3
was removed from the repo and added to .gitignore as per discussion on
GitHub
(https://github.com/ioccc-src/mkiocccentry/issues/338#issuecomment-1316142467).

This commit removes the hackish feeling from commit
916db953b9462f395e9be6775bb13ee6cfac5f63.

Notice that doing make dbg will (if everything is compiled) show:

    make[1]: Nothing to be done for `all'.

but whether that's a problem or not I do not know.

I believe this is good progress and with the way this is done similar
could be done for new dyn_array and jparse repos. There still should
probably be a Makefile for the test/ subdirectory though.