Make CFLAGS for dbg match our repo not dbg repo
It might be okay otherwise but in the past I've noticed problems mixing
CFLAGS in two different parts of a program (though in that case it was
shared libraries and the main program). But it seems better to match
CFLAGS anyway especially since we have developmental flags that will
change later on when we're sure everything is sound.

This was not done by changing the dbg/Makefile but rather the make
invocation in the Makefile like so:

	${MAKE} -C dbg CFLAGS="${CFLAGS}"