Reorganise test/ and related rules
This is a major update in that a lot of files have been renamed
(specifically test/ -> ioccc_test/) but also several files have changed.

The test/ subdirectory was renamed ioccc_test/ so that using make test
will work well with the Makefile under the ioccc_test/ subdirectory. To
make the test targets one uses make ioccc_test which will do just the
compilation and linking. Running make test will run the test suite. If
these two were not separated it would mean that compiling the test tools
would also run the test suite which is not always wanted (usually not
wanted but this way we can run make test and expect the same behaviour
from before).

Updated scripts to refer to ioccc_test/ instead of test/.

Updated Makefile. The only time we cd to ioccc_test is for when running
the test suite. The cd ioccc_test && ... part has been removed as the
subdirectory now has a Makefile.

An important note about the Makefile under ioccc_test is that it does
not have a make depend rule. Perhaps it should and maybe the dbg repo
should have that too.

Whereas before when everything was already compiled and linked running
make would print nothing now one will regrettably see:

    $ make
    make -C dbg CFLAGS="-std=gnu11 -O0 -g -pedantic -Wall -Wextra -Werror -ferror-limit=0 "
    make[1]: Nothing to be done for `all'.

but whether this is a real problem or not I don't know. If it is it
probably can be fixed but it's cleaner now that ioccc_test/ has a
Makefile.

A note is that the test_txzchk/bad error files had to be updated because
of the change in path. At the same time I updated the README under that
subdirectory.

The only subdirectory (currently) that needs a Makefile is soup but that
can come at another time. Possibly it will be today but it might not be.