Cook some soup :-)
Specifically add soup/Makefile and update Makefile to use it along the
lines of for ioccc_test/Makefile.

With this commit one will notice at times 'nothing to be done' but if
this is a problem it can be filtered out.

With this commit we no longer have a need in the Makefile to cd to
another directory and then compile code - the Makefile does it via
recursive calls to make with the -C option.