Fix obscure bug with bool typedef - #417
The problem as reported is that under some systems running make clobber
all test would result in a compilation error:

    In file included from jnum_test.c:26:0:
    ../dbg/dbg.h:55:23: error: two or more data types in declaration specifiers
     typedef unsigned char bool;
			   ^
    In file included from ../dyn_array.h:52:0,
		     from ../util.h:42,
		     from ../json_parse.h:26,
		     from jnum_test.c:33:
    ../dbg/dbg.h:55:1: warning: useless type name in empty declaration [enabled by default]
     typedef unsigned char bool;
     ^
    In file included from jnum_test.c:26:0:
    ../util.h:53:23: error: two or more data types in declaration specifiers
     typedef unsigned char bool;
			   ^
    In file included from ../json_parse.h:26:0,
		     from jnum_test.c:33:
    ../util.h:53:1: warning: useless type name in empty declaration [enabled by default]
     typedef unsigned char bool;
     ^
    make[1]: *** [jnum_test.o] Error 1
    make[1]: Leaving directory `/var/tmp/mkiocccentry/test_ioccc'
    make: *** [test] Error 2

This is because bool was already defined and so the 'typedef unsigned
char bool;' was a problem. The solution is in a number of files add an
ifdef guard so that bool is only defined once - either from #including
stdbool.h or from the typedef. Also as for util.h we have to move the
inclusion of dyn_array.h so that it's below the boolean define as
otherwise we will get another issue in some systems.

This commit involves changes to some files that will possibly require
changes elsewhere - iocccsize.h (which I won't touch normally) and dbg.h
(which I'll do a pull request in the dbg repo).