Move dyn_array code to dyn_array/ subdirectory
This commit is very similar to the commit dealing with moving the dbg
facility to the dbg/ subdirectory except that there are some additions
made to the Makefile for this commit and the next commit with jparse:

Add rules that match the dbg repo cloning etc. but for dyn_array and
jparse. Since these repos do not yet exist the rules are disabled (the
actual commands are commented out and a message is printed saying that
the rule is disabled until the external repo exists).

I updated version.h to change the path for the dbg.h and the
dyn_array.h.

I tested this under macOS and CentOS so hopefully everything is in the
clear. All files that refer to dyn_array.h (and dyn_array.c) have been
updated.