Move jparse to subdirectory and fix dyn_array bug
This commit involved a lot of changes and it's bound to have some things
that might need to be cleaned up or fixed - but not now. It works under
both linux and macOS running several different combinations of make all
test clobber and the bug_report.sh script (which also has some fixes).

What matter here is that the json parser is now in its own directory
jparse with its own Makefile. A few things to note. First is that I had
to explicitly disable some warnings - a warning which were already
needed before but in the case of CentOS I had to disable it via a
different option (it works under macOS also). This is because either the
scanner or parser generates a function that is not actually used.

I did NOT move jstrencode or jstrdecode to jparse. I'm not sure if these
should be moved over or not. The test scripts still remain in
test_ioccc/.

The man pages for run_flex.sh and run_bison.sh have not been moved to
jparse/ but they should be and will be in a future commit. This problem
might be there for other files including other man pages. If so these
will also be fixed later.

This commit also fixes a problem with make test compiling utf8_test.o:
namely that I forgot to update the test_ioccc/Makefile to refer to
../dyn_array as a subdirectory so the file dyn_array.o could not be
linked in as it was no longer that path. I probably did not have the
problem because the object file was still on my system - though that's
just a guess.

I had to update the run_flex.sh and run_bison.sh scripts to take a new
option which specifies the directory where the grammar file should be
in. Defaults to ".". This was needed for bug_report.sh.

A lot more things changed in this commit but it was all done in order to
move the json parser to its own directory.