Clear all known XXX issues
We were able to clear a number of old XXX issues as follows:

In `json_parse.c`:

* XXX - this table assumes we process on a byte basis - XXX
* XXX - consider an approach that allowed for smaller UTF-8 non-ASCII encoding - XXX

We do process on a byte basis, particularly on the JSON parser.
We simply note that we do this in a comment.

XXX Are there any other checks that have to be done ?

No.

* XXX json_conv_bool_str() calls json_conv_bool() which will warn if the

Yes it does.  This is not an XXX issue.

In `json_parse.h`:

* XXX - this table assumes we process on a byte basis - XXX
* XXX - consider an approach that allowed for smaller UTF-8 non-ASCII encoding - XXX

We do process on a byte basis, particularly on the JSON parser.
We simply note that we do this in a comment.

In `mkiocccentry.c`:

* XXX - update XXX's to show how past IOCCC winner can find their IOCCC winner handle - XXX

We added IOCCC_WINNER_HANDLE_READY, which is currently undefined.  '

We turned the XXX's above into TODO's.

This can be done once the IOCCC Winner Handle System, which depends
on the www.ioccc.org site rebuild, is ready.

In `mkiocccentry.h`:

* XXX - change to real URL when ready
* XXX - change to real URL when ready
* XXX - change to #define when submit server is ready

We turned the XXX's above into TODO's.

This can be done once the submit server is ready.

In `sanity.c`:

* XXX This function should probably be updated to search for at least jparse as well if not other utils too.

We don't think this is needed as we call the JSON parser via the API,
not via the command.

In `dbg/Makefile`:

* XXX - ###################################################################### - XXX #
* XXX - In 2024 we will will support only c17 so C_STD will become c-std=gnu17 - XXX #
* XXX - ###################################################################### - XXX #

We turned the XXX's above into TODO's.

This can be done in 2024.

Also NOTE: These TODO's might change once issue #422 is resolved.

* XXX - this will be the production release value - XXX

We turned the XXX's above into TODO's.

This can be done prior to the formal 1.0 release.

Also NOTE: These TODO's might change once issue #422 is resolved.

In `test_ioccc/Makefile`:

* XXX - ###################################################################### - XXX #
* XXX - In 2024 we will will support only c17 so C_STD will become c-std=gnu17 - XXX #
* XXX - ###################################################################### - XXX #

We turned the XXX's above into TODO's.

This can be done in 2024.

Also NOTE: These TODO's might change once issue #422 is resolved.

* XXX - this will be the production release value - XXX

We turned the XXX's above into TODO's.

This can be done prior to the formal 1.0 release.

Also NOTE: These TODO's might change once issue #422 is resolved.

In `run_usage.sh`:

* XXX However it still won't work right if all the usage strings are the same as

Changed the XXX to a NOTE.

In `soup/Makefile"`:

* XXX - ###################################################################### - XXX #
* XXX - In 2024 we will will support only c17 so C_STD will become c-std=gnu17 - XXX #
* XXX - ###################################################################### - XXX #

We turned the XXX's above into TODO's.

This can be done in 2024.

Also NOTE: These TODO's might change once issue #422 is resolved.

* XXX - this will be the production release value - XXX

We turned the XXX's above into TODO's.

This can be done prior to the formal 1.0 release.

Also NOTE: These TODO's might change once issue #422 is resolved.