Move hexval in util to json_util
Since hexval is strictly for JSON it belongs in json_util. A point of
interest is it's not actually used in json_util but rather json_parse so
why can't we have it in json_parse? We could but it would require
renaming either it or the variable in jencchk() in json_parse.c as if
the array was in json_parse.c it would be a name clash. Renaming the
unsigned int hexval probably should be done anyway because it shares the
name of an array but for now it's not done.