Fix make all not compiling all
This addresses one of the issues in issue #422 which I noted in comment
https://github.com/ioccc-src/mkiocccentry/issues/422#issuecomment-1327661223.
The problem was indeed (as I suspected) due to dependencies. The soup:
rule in the Makefile did not have any dependencies so running make all
would mean that more was to be done the next time round.