Add new man pages - mostly resolve #275
The man pages for jsemcgen.sh, jsemtblgen and all_ref.sh (under soup/)
have been added.

These can be improved a bit perhaps but it's probably fair to say that
they're mostly complete. They do not have examples. Probably the
description for each tool could be made better as well.

With the exception of finding a good man2html (or else deciding that we
don't want this) and formatting (and consistency - both of which I will
do later on) this should resolve issue #275.