Fix make depend bug if compilation fails
If a binary fails to compile make depend could fail. This might seem
like a good idea but the problem is sometimes this leads to being unable
to run make depend which is actually needed to successfully compile the
binary in the first place when actually running make depend would solve
the problem. In other words the make depend should not depend on having
to compile everything first because this can prevent correct and
successful compilation and in fact prevent one from doing anything at
all until the Makefile dependencies are manually updated which defeats
the purpose of having make depend.

The files changed in this commit are the main Makefile and
soup/Makefile; the dyn_array/Makefile and dbg/Makefile don't yet have a
depend rule.

To make it clear it means the make depend should not have 'all' as a
dependency. I am pointing this out for when dbg/Makefile (and the
external dbg repo Makefile) and the dyn_array/Makefile (and the future
dyn_array repo Makefile).