Update all_ref and jsem{tbl,c}gen man pages
Although they're not strictly correct just yet they are better th an
they were and I need to get the modified files out of the way but don't
want to stash them. The reason they're not completely correct is they
should not actually refer to the IOCCC at all but currently do. The
reason they should not refer to it is when there's a jparse repo (after
the IOCCCMOCK I will do this) they will also be there as they're not
IOCCC specific. However the way they're worded for now is still an
improvement.

I also want to clarify something I said in commit
0185b5f378bb67ed558c39a4e18bf8742ffa2c6e: make chkentry does of course
show 'is up to date' when it's compiled and nothing has been updated
that necessitates a recompilation. But make all, make dbg and make
dyn_array do not and so neither should make jparse which is (as of that
commit) a subdirectory (just like dbg and dyn_array).