Added jparse/test_jparse/sample_JSON
This is in anticipation for proper man pages for the three tools
mentioned in commit f8919d4005c81f775599765b594d4af189281167 as well as
(much down the road) a jparse repo. All this has is a json file which
will be used in a simpler tool for demonstration purposes.