Move jparse test files to jparse/test_jparse
This is a major update in that a lot of files have been moved and a fair
number have been updated as well.

A new Makefile has been added also - jparse/test_jparse/Makefile. The
make test rule runs successfully from the directory itself and running
from the root directory make test and make release also work. The
bug_report.sh script was updated to work as well.

If I missed any references to the files that were moved from test_ioccc/
to jparse/test_jparse they should be easy to update. Perhaps the
Makefile has some - I think I got the majority of them if not all of
them - but since the two rules and bug_report.sh work and running

    grep -Er 'test_ioccc/j' *

looks good as well I think everything should be in order. The only
matches are the legacy files that no longer should exist and certainly
would not exist anywhere under jparse/.