Link in dbg.a and not dbg.o
As we now have a dbg.a under dbg/ link that in instead as that was the
purpose of the library in the first place.

Also removed from OBJFILES dbg/dbg.o as dbg rule (which is run by make
all) will do make -C dbg which will build it so there's no need to have
it in OBJFILES. This means make depend was also run and dbg/dbg.o
dependencies have been removed (this seems better because it allows for
dbg/Makefile - which is also another repo - to control its dependencies
though right now that Makefile does not have a depend rule).