Move jparse code to jparse/
This is a major update with a lot of files being moved and many files
were also changed. There is some clean up required and some things still
need to be done but I wanted these changes in. See below.

New file is the jparse/Makefile. Running make depend in the root
directory will run make depend in jparse/.

It needs to be cleaned up! This commit is just so the changes are in. I
will do some clean up in the next commit. There are some other things
that have to be done yet. Amongst these things is to make jparse into a
static library. That can also come later. There are some rules in the
main Makefile that need to be cleaned up and some that need to be added
as well like for the man pages. These can also come in a future commit.

The bug_report.sh was updated not only for this change but also to fix a
recent commit when dyn_test was moved from test_ioccc/ to dyn_array/.

I added to options in jparse.l noinput so that at least one warning was
no longer triggered. It might be we can remove the -Wno- warnings in the
jparse Makefile but I have to look into this further again at a later
time.