Minor fix in Makefile
In the previous commit I had to add to make depend a dependency for
jparse but I ended up putting in for dbg and dyn_array. These are not
needed. The reason jparse is needed is because the jparse rule will
generate files that are needed for make depend. I'm not sure if there
are any implications of if for instance make jparse fails for some
reason but nevertheless it's necessary for it to be there (maybe the
dependency should not be jparse but something else but this not known
yet).