Fixes for jparse/Makefile commit
This commit fixes several issues in
fd53b2e55486132e4be9de86daf1def558eba094 but there are still things that
need to be done which I will note after the changes.

There was a make depend issue (or maybe issues) where amongst other
things at least one object file had dependencies listed twice.

A number of rules were removed from the main Makefile that are dealt
with in jparse/Makefile. Perhaps more can be removed but I'm not certain
of this and right now it's more than I can deal with. One that has to
remain in targets, it seems, is jnum_gen. This seemed to be a problem
when running bug_report.sh even though it it's in TARGETS in the
jparse/Makefile.

I changed some (more needs to be done) of the order of things to be more
consistent with the other Makefiles. I should say much more needs to be
done with the ordering.

What is not done is probably more than what is done but given that we're
not sure if the way the multi Makefile system is set up is what we want
in the end and given that I have some other things going on I will not
worry about it at least for now. These things are certainly not done:

- Building of a static library. This would not take much effort and it
can be similar to the dyn_array and dbg libraries. However right now I
cannot deal with this so it'll have to be pushed out.

- Making the Makefile more consistent with the other Makefiles. Given
that these are themselves subject to change and given some things going
on elsewhere I am not going to try doing this for now.

- Not all Makefiles have depend rules. Those that compile anything
should certainly have depend rules.

- Should jparse/Makefile have a test rule that runs jparse_test.sh?
Probably but it does not currently have this.

- The man pages although installed from jparse/man -> man/ should be
done better namely it does not quite match the other Makefiles. However
as I noted in comments maybe those could be improved too as running make
repeatedly will still give you:

    make -C man all
    ../dbg/dbg.3 -> dbg.3
    make[2]: Nothing to be done for `dyn_array.3'.
    ../dyn_array/dyn_array.3 -> dyn_array.3
    make[2]: Nothing to be done for `man/jparse.1'.
    ../jparse/man/jparse.1 -> jparse.1
    make[2]: Nothing to be done for `man/jparse_test.8'.
    ../jparse/man/jparse_test.8 -> jparse_test.8

(When I say I did not match it I mean variables and set up in the
Makefile - the critical rules are there however.)

But at least with these commits the jparse code and data are now under
jparse/. The required clean up (and I'm not even sure of everything that
needs to be done there) and the building of the static library can be
done and implemented now that the jparse/Makefile works. Running
bug_report.sh reports no problems and running make prep reports no
problems either.